requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.4.2
SQLAlchemy>=2.0.29
aiomysql>=0.2.0
pymysql>=1.0.0
greenlet>=1.1.0
python-dotenv>=1.0.1
APScheduler>=3.10.4
scikit-learn>=1.3.2
numpy>=1.26.4
gensim>=4.3.2
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2
python-multipart>=0.0.6
fastapi-users[sqlalchemy]>=13.0.0
alembic>=1.13.1
pytest>=8.1.1
httpx>=0.27.0
asyncpg>=0.29.0
requests>=2.31.0
scipy>=1.13.0 